fix(indexer): compute direct-K K-page byte offsets in Int64
The warp-owns-tokens direct-L2 score path (39d7404) computes the K
fragment byte offset as pid * Int32(k_quant_page_stride). Serving
integrations pass strided views of a packed per-page record (observed
page stride 1,077,120 B, vs 8448 for a bare [pages,64,128]+scales
tensor), so the Int32 product wraps at pid >= ceil(2^31 / 1,077,120) =
1994; the wrapped offset sign-extends and the ld.global.nc lands ~1.2 GB
before the tensor: CUDA_ERROR_ILLEGAL_ADDRESS on the first long-context
request of a DSV4-Flash DSpark deployment (GB10 takes this path for
every verify batch at rows >= 3; any SM count reaches it once
ctas_per_group <= 16).

Compute the page term as Int64(pid) * Int64(stride) and carry the
offset through _score_tokens_direct_k as Int64, matching the staged
pipeline's Int64(page_id) * page_stride_bytes convention in
_load_permute_k_page_g2s / _stream_issue_k_page_cp_async. Intra-page
terms stay Int32 and are widened for the add.

Caught via GPU core dump on a 2-node GB10 TP2 serving pair: Warp MMU
Fault in SparseNSAFusedIndexerKernel at IMAD R6,R41,0x106f80,R6 with
R41 = pid = 6839. The regression test reconstructs the packed stride
with pid >= 2000 and fails on the unfixed kernel in ~4 s (verified on
SM120/RTX PRO 6000 and SM121/GB10); full fused-indexer suite passes.

+++ b12x/attention/indexer/fused_indexer.py
@@ -350,7 +350,7 @@ def _score_tokens_direct_k(
     s_w: cute.Tensor,
     num_heads: Int32,
     k_quant_bytes: cute.Tensor,
-    k_byte_off: Int32,
+    k_byte_off: Int64,
     lane: Int32,
     num_q_head_tiles: cutlass.Constexpr[int],
 ):
@@ -368,13 +368,13 @@ def _score_tokens_direct_k(
     q_row = lane & Int32(15)
     q_half = lane >> Int32(4)
     k0_0 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off))
-    k1_0 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(16)))
-    k0_1 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(32)))
-    k1_1 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(48)))
-    k0_2 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(64)))
-    k1_2 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(80)))
-    k0_3 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(96)))
-    k1_3 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int32(112)))
+    k1_0 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(16)))
+    k0_1 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(32)))
+    k1_1 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(48)))
+    k0_2 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(64)))
+    k1_2 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(80)))
+    k0_3 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(96)))
+    k1_3 = ld_global_nc_u32(get_ptr_as_int64(k_quant_bytes, k_byte_off + Int64(112)))
     total0 = Float32(0.0)
     total1 = Float32(0.0)
     qgrp = lane // Int32(4)
@@ -1513,9 +1513,9 @@ class SparseNSAFusedIndexerKernel:
                 t1 = Float32(0.0)
                 if pid >= Int32(0):
                     k_off = (
-                        pid * Int32(self.k_quant_page_stride)
-                        + (tip0 + lane // Int32(4)) * Int32(_INDEX_HEAD_DIM)
-                        + lane4 * Int32(4)
+                        Int64(pid) * Int64(self.k_quant_page_stride)
+                        + Int64((tip0 + lane // Int32(4)) * Int32(_INDEX_HEAD_DIM))
+                        + Int64(lane4 * Int32(4))
                     )
                     t0, t1 = _score_tokens_direct_k(
                         q_smem_base_addr,
